refactor(rotate_polytopes): split into per-mode modules (state, composer, filmstrip, active, shapes, ui)
